+++ R/lib.R
@@ -9,77 +9,64 @@ ping_ollama <- function(server = NULL, silent = FALSE) {
 
   if (is.null(server)) server <- getOption("rollama_server",
                                            default = "http://localhost:11434")
-  res <- try({
-    httr2::request(server) |>
-      httr2::req_url_path("api/version") |>
-      httr2::req_perform() |>
-      httr2::resp_body_json()
-  }, silent = TRUE)
 
-  if (!methods::is(res, "try-error") & purrr::pluck_exists(res, "version")) {
-    if (!silent) cli::cli_inform(
-      "{cli::col_green(cli::symbol$play)} Ollama (v{res$version}) is running at {.url {server}}!"
-    )
-    out <- TRUE
-    attr(out, "v") <- res$version
-    invisible(out)
-  } else {
-    if (!silent) {
-      cli::cli_alert_danger("Could not connect to Ollama at {.url {server}}")
-      print(res)
+  out <- purrr::map_lgl(server, function(sv) {
+    res <- try({
+      httr2::request(sv) |>
+        httr2::req_url_path("api/version") |>
+        httr2::req_perform() |>
+        httr2::resp_body_json()
+    }, silent = TRUE)
+
+    if (!methods::is(res, "try-error") && purrr::pluck_exists(res, "version")) {
+      if (!silent) cli::cli_inform(
+        "{cli::col_green(cli::symbol$play)} Ollama (v{res$version}) is running at {.url {sv}}!"
+      )
+      return(TRUE)
+    } else {
+      if (!silent) {
+        cli::cli_alert_danger("Could not connect to Ollama at {.url {sv}}")
+      }
+      return(FALSE)
     }
-    invisible(FALSE)
-  }
-
+  })
+  invisible(all(out))
 }
 
 
-build_req <- function(model, msg, server, images, model_params, format, template) {
+build_req <- function(model,
+                      msg,
+                      server,
+                      images,
+                      model_params,
+                      format,
+                      template) {
 
-  if (is.null(model)) model <- getOption("rollama_model", default = "llama3")
+  if (is.null(model)) model <- getOption("rollama_model", default = "llama3.1")
   if (is.null(server)) server <- getOption("rollama_server",
                                            default = "http://localhost:11434")
-  check_model_installed(model)
-  req_data <- purrr::map(model, function(m) {
-    list(model = m,
-         messages = msg,
-         stream = FALSE,
-         options = model_params,
-         format = format,
-         template = template) |>
-      purrr::compact() |> # remove NULL values
-      make_req(server = server,
-               endpoint = "/api/chat",
-               perform = FALSE)
-  })
-
-  if (getOption("rollama_verbose", default = interactive())) {
-    cli::cli_progress_step("{model} {?is/are} thinking {cli::pb_spin}")
-
-    rp <- purrr::map(req_data, function(r) callr::r_bg(httr2::req_perform,
-                                                       args = list(req = r),
-                                                       package = TRUE))
-
-    while (any(purrr::map_lgl(rp, function(r) r$is_alive()))) {
-      cli::cli_progress_update()
-      Sys.sleep(2 / 100)
-    }
-    resp <- purrr::map(rp, function(r) r$get_result()) |>
-      purrr::map(httr2::resp_body_json)
-    cli::cli_progress_done()
-  } else {
-    resp <- purrr::map(req_data, httr2::req_perform) |>
-      purrr::map(httr2::resp_body_json)
+  seed <- getOption("rollama_seed")
+  if (!is.null(seed) && !purrr::pluck_exists(model_params, "seed")) {
+    model_params <- append(model_params, list(seed = seed))
   }
+  check_model_installed(model, server = server)
+  req_data <- purrr::map(msg, function(ms) {
+    purrr::map(model, function(m) {
+      list(model = m,
+           messages = ms,
+           stream = FALSE,
+           options = model_params,
+           format = format,
+           template = template) |>
+        purrr::compact() |> # remove NULL values
+        make_req(server = sample(server, 1, prob = as_prob(names(server))),
+                 endpoint = "/api/chat",
+                 perform = FALSE)
+    })
+  }) |>
+    unlist(recursive = FALSE)
 
-  es <- purrr::map(resp, "error") |>
-    unlist()
-  if (length(es) > 0) {
-    names(es) <- rep("!", length(es))
-    cli::cli_abort(es)
-  }
-
-  return(resp)
+  return(req_data)
 }
 
 
@@ -89,6 +76,9 @@ make_req <- function(req_data, server, endpoint, perform = TRUE) {
     httr2::req_body_json(prep_req_data(req_data), auto_unbox = FALSE) |>
     # turn off errors since error messages can't be seen in sub-process
     httr2::req_error(is_error = function(resp) FALSE) |>
+    # see https://github.com/JBGruber/rollama/issues/23
+    httr2::req_options(timeout_ms = 1000 * 60 * 60 * 24,
+                       connecttimeout_ms = 1000 * 60 * 60 * 24) |>
     httr2::req_headers(!!!get_headers())
   if (perform) {
     r <- r |>
@@ -96,6 +86,75 @@ make_req <- function(req_data, server, endpoint, perform = TRUE) {
       httr2::resp_body_json()
   }
   return(r)
+}
+
+
+perform_reqs <- function(reqs, verbose) {
+
+  model <- purrr::map_chr(reqs, c("body", "data", "model")) |>
+    unique()
+  pb <- FALSE
+  if (verbose) {
+    pb <- list(
+      clear = TRUE,
+      format = c("{cli::pb_spin} {getOption('model')} {?is/are} thinking about ",
+                 "{cli::pb_total - cli::pb_current}/{cli::pb_total} question{?s}")
+    )
+  }
+
+  withr::with_options(list(cli.progress_show_after = 0, model = model), {
+    resps <- httr2::req_perform_parallel(
+      reqs = reqs,
+      on_error = "continue",
+      progress = pb
+    )
+  })
+
+  fails <- httr2::resps_failures(resps) |>
+    purrr::map_chr("message")
+
+  # all fails
+  if (length(fails) == length(reqs)) {
+    cli::cli_abort(fails)
+  } else if (length(fails) < length(reqs) && length(fails) > 0) {
+    error_counts <- table(fails)
+    for (f in names(error_counts)) {
+      if (error_counts[f] > 2) {
+        cli::cli_alert_danger("error ({error_counts[f]} times): {f}")
+      } else {
+        cli::cli_alert_danger("error: {f}")
+      }
+    }
+    for (f in fails) {
+      cli::cli_alert_danger(f)
+    }
+  }
+
+  httr2::resps_successes(resps)
+}
+
+
+perform_req <- function(reqs, verbose) {
+
+  if (verbose) {
+    model <- purrr::map_chr(reqs, c("body", "data", "model")) |>
+      unique()
+
+    id <- cli::cli_progress_bar(format = "{cli::pb_spin} {model} {?is/are} thinking",
+                                clear = TRUE)
+
+    rp <- callr::r_bg(httr2::req_perform,
+                      args = list(req = reqs[[1]]),
+                      package = TRUE)
+
+    while (rp$is_alive()) {
+      cli::cli_progress_update(id = id)
+      Sys.sleep(2 / 100)
+    }
+    return(list(rp$get_result()))
+  }
+
+  list(httr2::req_perform(reqs[[1]]))
 }
 
 
@@ -167,11 +226,13 @@ pgrs <- function(resp) {
         paste(round(the$str_prgs$done / the$str_prgs$total * 100, 0), "%")
       if (the$str_prgs$done != the$str_prgs$total) {
         the$str_prgs$speed <-
-          prettyunits::pretty_bytes(
-            the$str_prgs$done /
-              (as.integer(Sys.time()) - as.integer(the$str_prgs$pb_start))
-          )
-      } else the$str_prgs$speed <- 1L
+          the$str_prgs$done /
+          (as.integer(Sys.time()) - as.integer(the$str_prgs$pb_start))
+        if (is.numeric(the$str_prgs$speed))
+          the$str_prgs$speed <- prettyunits::pretty_bytes(the$str_prgs$speed)
+      } else {
+        the$str_prgs$speed <- 1L
+      }
 
       if (!isTRUE(the$str_prgs$pb == the$str_prgs$f)) {
         cli::cli_progress_bar(