+++ R/models.r
@@ -8,13 +8,13 @@
 #'
 #' **Model names**: Model names follow a model:tag format, where model can have
 #' an optional namespace such as example/model. Some examples are
-#' orca-mini:3b-q4_1 and llama3:70b. The tag is optional and, if not provided,
+#' orca-mini:3b-q4_1 and llama3.1:70b. The tag is optional and, if not provided,
 #' will default to latest. The tag is used to identify a specific version.
 #'
-#' @param model name of the model. Defaults to "llama3" when `NULL` (except in
-#'   `delete_model`).
+#' @param model name of the model(s). Defaults to "llama3.1" when `NULL` (except
+#'   in `delete_model`).
 #' @param insecure allow insecure connections to the library. Only use this if
-#'   you are pulling from your own library during development. description
+#'   you are pulling from your own library during development.
 #' @param destination name of the copied model.
 #' @inheritParams query
 #'
@@ -24,22 +24,38 @@
 #'
 #' @examples
 #' \dontrun{
+#' # download a model and save information in an object
 #' model_info <- pull_model("mixtral")
 #' # after you pull, you can get the same information with:
 #' model_info <- show_model("mixtral")
+#' # pulling models from Hugging Face Hub is also possible
+#' pull_model("https://huggingface.co/oxyapi/oxy-1-small-GGUF:Q2_K")
 #' }
-pull_model <- function(model = NULL, server = NULL, insecure = FALSE) {
+pull_model <- function(model = NULL,
+                       server = NULL,
+                       insecure = FALSE,
+                       verbose = getOption("rollama_verbose",
+                                           default = interactive())) {
 
-  if (is.null(model)) model <- getOption("rollama_model", default = "llama3")
+  if (is.null(model)) model <- getOption("rollama_model", default = "llama3.1")
   if (is.null(server)) server <- getOption("rollama_server",
                                            default = "http://localhost:11434")
 
-  httr2::request(server) |>
-    httr2::req_url_path_append("/api/pull") |>
-    httr2::req_body_json(list(name = model, insecure = insecure)) |>
-    httr2::req_perform_stream(callback = pgrs, buffer_kb = 0.1)
+  if (length(model) > 1L) {
+    for (m in model) pull_model(m, server, insecure, verbose)
+  }
 
-  cli::cli_process_done(.envir = the)
+  # flush progress
+  the$str_prgs <- NULL
+  req <- httr2::request(server) |>
+    httr2::req_url_path_append("/api/pull") |>
+    httr2::req_body_json(list(name = model, insecure = insecure))
+  if (verbose) {
+    httr2::req_perform_stream(req, callback = pgrs, buffer_kb = 0.1)
+    cli::cli_process_done(.envir = the)
+  } else {
+    httr2::req_perform(req)
+  }
   cli::cli_alert_success("model {model} pulled succesfully")
   the$str_prgs <- NULL
 
@@ -51,7 +67,7 @@ pull_model <- function(model = NULL, server = NULL, insecure = FALSE) {
 #' @export
 show_model <- function(model = NULL, server = NULL) {
 
-  if (is.null(model)) model <- getOption("rollama_model", default = "llama3")
+  if (is.null(model)) model <- getOption("rollama_model", default = "llama3.1")
   if (is.null(server)) server <- getOption("rollama_server",
                                            default = "http://localhost:11434")
   if (length(model) != 1L) cli::cli_abort("model needs to be one model name.")
@@ -63,8 +79,7 @@ show_model <- function(model = NULL, server = NULL) {
     httr2::req_perform() |>
     httr2::resp_body_json() |>
     purrr::list_flatten(name_spec = "{inner}") |>
-    purrr::compact() |>
-    tibble::as_tibble()
+    as_tibble_onerow()
 }
 
 
@@ -90,7 +105,7 @@ show_model <- function(model = NULL, server = NULL) {
 #' @examples
 #' modelfile <- system.file("extdata", "modelfile.txt", package = "rollama")
 #' \dontrun{create_model("mario", modelfile)}
-#' modelfile <- "FROM llama3\nSYSTEM You are mario from Super Mario Bros."
+#' modelfile <- "FROM llama3.1\nSYSTEM You are mario from Super Mario Bros."
 #' \dontrun{create_model("mario", modelfile)}
 create_model <- function(model, modelfile, server = NULL) {
 
@@ -102,6 +117,8 @@ create_model <- function(model, modelfile, server = NULL) {
     modelfile <- paste0(modelfile, collapse = "\n")
   }
 
+  # flush progress
+  the$str_prgs <- NULL
   httr2::request(server) |>
     httr2::req_url_path_append("/api/create") |>
     httr2::req_method("POST") |>