+++ NEWS.md
@@ -1,3 +1,16 @@
+# rollama 0.2.0
+
+* added make_query() function to facilitate easier annotation
+* added more output formats to query()/chat()
+* improved performance of embed_text()
+* improved performance of query() for multiple queries
+* changed default model to llama3.1
+* added option to employ multiple servers
+* pull_model() gained verbose option
+* improved annotation vignette 
+* added vignette on how to use Hugging Face Hub models
+* some bug fixes
+
 # rollama 0.1.0
 
 * adds function `check_model_installed`